+++ src/main/java/com/tsi/lerai/foulkes/program/ActorRepo.java
@@ -2,6 +2,6 @@ package com.tsi.lerai.foulkes.program;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ActorRepo extends CrudRepository<Actor,Integer> {
+public interface ActorRepo extends CrudRepository<Actor,Long> {
 
 }
